#163 [REFACTOR] Public 폴더 내부 이미지 및 아이콘 정리
