+++ components/Project/Management/CreateProjectCard.tsx
@@ -2,7 +2,7 @@ import {
   ProjectCardCommon,
   ProjectCardCommonProps,
 } from 'components/Project/Management/ProjectCardCommon';
-import PluseIconSVG from 'public/images/plus_icon.svg';
+import PluseIconSVG from 'public/icons/plus.svg';
 
 export const CreateProjectCard = (props: ProjectCardCommonProps) => {
   return <CreateProjectCardView {...props} />;
