+++ components/ShoppingMallReview/ComprehensiveRates.tsx
@@ -2,8 +2,7 @@ import Image from 'next/image';
 import React from 'react';
 
 import risingArray from 'public/images/rising_array.png';
-import grayStar from 'public/images/gray_star.png';
-import emptyStar from 'public/images/empty_star.png';
+import grayStar from 'public/images/empty_star_img.png';
 import * as S from './ComprehensiveRates.styles';
 import { ShoppingMallReviewInfo } from 'typings/reviews';
