+++ components/common/Oauth/KakaoOauth.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import styled from '@emotion/styled';
 
-import KakaoIcon from 'public/images/kakao_icon.svg';
+import KakaoIcon from 'public/icons/kakao.svg';
 
 declare global {
   interface Window {
